requirements.txt
pystray
dxcam
pyyaml
flask
requests
Pillow
numpy
qoi
pywin32
pyinstaller
appdirs
ffmpeg-python
flask_restful
pynvml
winshell
Windows-Toasts
lovely_logger